requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
setuptools==70.1.0
SQLAlchemy==1.4.29
typing_extensions==4.12.2
Werkzeug==2.0.3
pycodestyle